pyproject.toml
[project]
name = "hotel-captions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=3.2.0",
    "fastapi>=0.115.7",
    "imagehash>=4.3.1",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.0",
    "numpy>=2.2.2",
    "onnxruntime>=1.20.1",
    "optimum>=1.23.3",
    "pillow>=11.1.0",
    "protobuf>=5.29.3",
    "requests>=2.32.3",
    "ruff>=0.9.2",
    "scikit-learn>=1.6.1",
    "sentencepiece>=0.2.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.48.1",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
